fix: include `tier` in names of production workloads
Previously, there was a half-thought-out attempt at omitting "production-y"
tiers from resource names generated by the `defaults` module.

The control mechanism for this was not very clear, so this PR replaces
it with a simple boolean toggle, which defaults to `true` (i.e. tiers
are always included in names by default).